Added unit-tests for Services/Classes/AccessGroupService.cs

+++ AccessManagementApp/Services/Classes/AccessGroupService.cs
@@ -18,6 +18,15 @@ namespace AccessManagementApp.Services.Classes
         }
         public async Task<AccessGroupModel> Create(CreateAccessGroupModel item)
         {
+            if(item == null)
+                throw new ArgumentNullException(nameof(item));
+
+            if(string.IsNullOrEmpty(item.Name))
+                throw new ArgumentNullException(nameof(item.Name));
+
+            if (item.Accesses == null || item.Accesses.Count == 0)
+                throw new ArgumentNullException(nameof(item.Accesses));
+
             var accessGroup = _mapper.Map<AccessGroup>(item);
 
             var createdAccessGroup = await _accessGroupRepository.Create(accessGroup);
@@ -29,6 +38,11 @@ namespace AccessManagementApp.Services.Classes
 
         public async Task<AccessGroup> Delete(int id)
         {
+            var itemToDelete = await _accessGroupRepository.GetById(id);
+
+            if (itemToDelete == null)
+                throw new KeyNotFoundException("Unable to delete non-existing item");
+
             var deletedAccessGroup = _mapper.Map<AccessGroup>(await _accessGroupRepository.Delete(id));
 
             return deletedAccessGroup;
@@ -57,6 +71,20 @@ namespace AccessManagementApp.Services.Classes
 
         public async Task<AccessGroupModel> Update(UpdateAccessGroupModel item)
         {
+            if (item == null)
+                throw new ArgumentNullException(nameof(item));
+
+            if (string.IsNullOrEmpty(item.Name))
+                throw new ArgumentNullException(nameof(item.Name));
+
+            if (item.Accesses == null || item.Accesses.Count == 0)
+                throw new ArgumentNullException(nameof(item.Accesses));
+
+            var itemToUpdate = await _accessGroupRepository.GetById(item.Id);
+
+            if (itemToUpdate == null)
+                throw new KeyNotFoundException("Unable to edit non-existing item");
+
             var accessGroup = _mapper.Map<AccessGroup>(item);
 
             var updatedAccessGroup = await _accessGroupRepository.Update(accessGroup);
